Finished the Slack Automation

+++ src/types/slack/internal/MongoDBInstallationStore.mts
@@ -6,6 +6,7 @@ export default class MongoDBInstallationStore implements InstallationStore {
         if (installation.team === undefined) throw new Error("Team is required");
         await InstallationStoreSchema.findOneAndUpdate({
             teamId: installation.team.id,
+            userId: installation.user.id,
         }, {
             userId: installation.user.id,
             teamId: installation.team.id,
@@ -19,12 +20,35 @@ export default class MongoDBInstallationStore implements InstallationStore {
     }
 
     async fetchInstallation(query: InstallationQuery<boolean>): Promise<Installation<"v1" | "v2", boolean>> {
-        const installation = await InstallationStoreSchema.findOne({ teamId: query.teamId });
-        if (installation === null) throw new Error("Installation not found");
+        let installation = await InstallationStoreSchema.findOne({
+            teamId: query.teamId,
+            userId: query.userId,
+            installation: {
+                $exists: true,
+                $ne: null,     
+            },
+            $expr: {
+                $eq: ['$installation.isEnterpriseInstall', query.isEnterpriseInstall],
+            }
+        });
+        if (!installation && query.teamId) {
+            installation = await InstallationStoreSchema.findOne({
+                teamId: query.teamId,
+                installation: {
+                    $exists: true,
+                    $ne: null,     
+                },
+                $expr: {
+                    $eq: ['$installation.isEnterpriseInstall', query.isEnterpriseInstall],
+                }
+            });
+            installation!.installation.user = null as any;
+        }
+        if (!installation) return null as any;
         return installation.installation;
     }
 
     async deleteInstallation(query: InstallationQuery<boolean>): Promise<void> {
-        await InstallationStoreSchema.deleteOne({ teamId: query.teamId });
+        await InstallationStoreSchema.deleteOne({ teamId: query.teamId, userId: query.userId });
     }
 }
